backend/requirements.txt
flask
flask-cors
plaid
python-dotenv
requests
pymongo
pandas
scikit-learn
shap
secure-smtplib